Don't cleanup tests when TEST_NOCLEANUP is set
This let us debug the state of the test after it fails easily.

+++ test/pkg/gitea/setup.go
@@ -8,12 +8,13 @@ import (
 	"testing"
 
 	"github.com/google/go-github/v53/github"
+	"gotest.tools/v3/assert"
+
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/params"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/params/info"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/provider/gitea"
 	"github.com/openshift-pipelines/pipelines-as-code/test/pkg/options"
 	"github.com/openshift-pipelines/pipelines-as-code/test/pkg/repository"
-	"gotest.tools/v3/assert"
 )
 
 func CreateProvider(ctx context.Context, giteaURL, user, password string) (gitea.Provider, error) {
@@ -76,6 +77,10 @@ func Setup(ctx context.Context) (*params.Run, options.E2E, gitea.Provider, error
 }
 
 func TearDown(ctx context.Context, t *testing.T, topts *TestOpts) {
+	if os.Getenv("TEST_NOCLEANUP") == "true" {
+		topts.ParamsRun.Clients.Log.Infof("Not cleaning up and closing PR since TEST_NOCLEANUP is set")
+		return
+	}
 	repository.NSTearDown(ctx, t, topts.ParamsRun, topts.TargetNS)
 	_, err := topts.GiteaCNX.Client.DeleteRepo(topts.Opts.Organization, topts.TargetNS)
 	assert.NilError(t, err)
